refactor: ConfigMapSecret and EnvironmentOverride components - streamline state management and improve data fetching with useQuery
