+++ src/components/ciConfig/CIConfig.tsx
@@ -19,7 +19,7 @@ import { showError, Progressing } from '@devtron-labs/devtron-fe-common-lib'
 import { getGitProviderIcon, sortObjectArrayAlphabetically } from '../common'
 import { getDockerRegistryMinAuth } from './service'
 import { getSourceConfig, getCIConfig } from '../../services/service'
-import { ComponentStates } from '../../Pages/Shared/EnvironmentOverride/EnvironmentOverrides.types'
+import { ComponentStates } from '@Components/CIPipelineN/types'
 import { CIConfigProps, MaterialOptionType } from './types'
 import './CIConfig.scss'
 import CIConfigForm from './CIConfigForm'
