+++ src/components/ciConfig/types.tsx
@@ -32,10 +32,9 @@ import {
     WorkflowType,
 } from '@devtron-labs/devtron-fe-common-lib'
 
-import { EnvironmentWithSelectPickerType } from '@Components/CIPipelineN/types'
+import { ComponentStates, EnvironmentWithSelectPickerType } from '@Components/CIPipelineN/types'
 import { OptionTypeWithIcon } from '@Components/externalLinks/ExternalLinks.type'
 
-import { ComponentStates } from '../../Pages/Shared/EnvironmentOverride/EnvironmentOverrides.types'
 import { ConfigOverrideWorkflowDetails } from '../../services/service.types'
 import { CiPipelineResult } from '../app/details/triggerView/types'
 import { CIPipelineDataType } from '../ciPipeline/types'
